# Optimize the uniqueTags function to be faster with less allocs