This change optimizes the internal `uniqeTags()` function to be more efficient.
This was done with three modifications being made:

* use `copy()` instead of `append()`
* when allocating the map, ensure it's large enough to avoid further allocations
* for the map, use values with type `struct{}` instead of `bool`

The changes above resulted in the following improvements:

```
benchmark                 old ns/op     new ns/op     delta
BenchmarkUniqueTags-8     12692         6619          -47.85%

benchmark                 old allocs     new allocs     delta
BenchmarkUniqueTags-8     8              3              -62.50%

benchmark                 old bytes     new bytes     delta
BenchmarkUniqueTags-8     6805          4152          -38.99%
```

Signed-off-by: Tim Heckman <t@heckman.io>

/cc @akrylysov @deepakprabhakara